drivers/flash: Preliminary driver for flash_area device
This is attempt to move flash map functionality into flash
device API.  This way flash API could be directly used to
operate on flash areas defined as fixed-partitions.
The flash map API would have to partially be left as it is required
by external projects like mcumgr or mcuboot.
Nevertheless it may be significantly reduced and some operations
in other parts of Zephyr, like flash_img API, can be ported
to work on flash devices directly.

Signed-off-by: Dominik Ermel <dominik.ermel@nordicsemi.no>