tests/fs/littlefs: Allow any combination of partitions
The commit changes meaning of CONFIG_APP_TEST_CUSTOM for the
test as previously it would add medium and large partition to
test setup, and additionally do custom testing for small partition,
now it only adds the custom testing of small partition.
The test has been changed to allow any combination of partitions
to exist on a device; this allows to run test on devices that are
not able to define all required partitions at once, due to flash size,
but can afford to define combination of them or single partition
at once.

Signed-off-by: Dominik Ermel <dominik.ermel@nordicsemi.no>